+++ mmdet/models/necks/__init__.py
@@ -12,7 +12,7 @@ from .pafpn import PAFPN
 from .rfp import RFP
 from .ssd_neck import SSDNeck
 from .yolo_neck import YOLOV3Neck
-from .yolo_pafpn import YOLOXPAFPN
+from .yolox_pafpn import YOLOXPAFPN
 
 __all__ = [
     'FPN', 'BFP', 'ChannelMapper', 'HRFPN', 'NASFPN', 'FPN_CARAFE', 'PAFPN',
